Replace DiscordSocketClient with adapter to improve testability
The dependency in the DiscordBot class on DiscordSocketClient has been replaced with an interface IDiscordSocketClientAdapter and its implementation DiscordSocketClientAdapter. This will improve the testability of the class as we can now use a mock or stub for IDiscordSocketClientAdapter in tests. The adapter also encapsulates the handling of events from DiscordSocketClient, enhancing the separation of concerns in our code. DIRegistrations.cs  has been updated accordingly to provide an instance of DiscordSocketClientAdapter instead of DiscordSocketClient.