Move InteractionService to a DiscordBot field for testing.
The DiscordBot class now includes an instance of IInteractionServiceAdapter named _interactionService. This change was necessitated by the need to use an interaction service that can handle command centers, such as adding new modules dynamically, within the Discord bot. Additionally, an overloaded constructor was added for testing purposes, to allow injection of an InteractionService mock. Two new classes - IInteractionServiceAdapter and InteractionServiceAdapter - were created to facilitate this. These changes aim to improve extensibility and testability.