Rename GoogleCustomSearchService to SteamSearchService
The GoogleCustomSearchService class has been renamed to SteamSearchService. This new name reflects that the service is now specifically for Steam searches, making it more understandable for future developers working with this code. This was applied across all involved files, adapting variables, comments, and dependencies to match the new service name. An additional function `GetAppAsync()` was also implemented in the `ISteamStoreService` interface.