Rename service and update documentation details
Services and test files have been renamed for clarity. The name of the `SteamStoreServices` class has been changed to `SteamStoreService` and it is reflected across the entire project. This change standardizes the service naming convention. Furthermore, several documentation lines were updated to accurately describe the methods and classes they are attached to.