Rename SteamSearchService to SteamStoreSearchService and refactor
The rename clarifies the use of the service as it specifically searches the Steam Store and not a general Steam search. This change has affected various classes where SteamSearchService was used or tested. Other refactoring has been carried out to improve readability and performance. The related interface has also been renamed to match the updated service.